requirements.txt
aioconsole==0.6.1
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
Pillow==9.5.0
requests==2.31.0
six==1.16.0
urllib3==2.0.3
watchdog==3.0.0
opencv-python
pytesseract
pystray==0.19.4
# 高精度OCR用追加ライブラリ（オプション）
easyocr
google-cloud-vision
